src/requirements.txt
jsonpickle
websocket-client
algorum-quant-client-py3~=1.0.24